// templates/salesforce/types.ts
interface Attributes {
    type: string;
    url: string;
}

export interface ValidationRecord {
    attributes: Attributes;
    Id: string;
    ValidationName: string;
    Metadata: Metadata;
}

export interface ValidationRuleResponse {
    size: number;
    totalSize: number;
    done: boolean;
    queryLocator: string | null;
    entityTypeName: string;
    records: ValidationRecord[];
}

export interface SalesforceAccount {
    attributes: Attributes;
    Id: string;
    Name: string;
    Description: string | null;
    Website: string | null;
    Industry: string | null;
    BillingCity: string | null;
    BillingCountry: string | null;
    OwnerId: string;
    Owner: NamedEntity;
    NumberOfEmployees: number | null;
    LastModifiedDate: string;
}
// ...
interface NamedEntity {
    attributes: Attributes;
    Name: string;
}